pyproject.toml
[project]
name = "deliberative-mas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
